+++ module-api/src/test/java/com/junior/controller/qna/QnaUserControllerTest.java
@@ -1,4 +1,4 @@
-package com.junior.controller;
+package com.junior.controller.qna;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.junior.config.SecurityConfig;
